requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
pypdf
sentence_transformers
langchain_huggingface
faiss-cpu
hf_xet
chromadb
langchain_chroma
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette